+++ server/getObjectByPath.go
@@ -16,7 +16,6 @@ import (
 )
 
 func (h AppServer) getObjectByPath(ctx context.Context, w http.ResponseWriter, r *http.Request) *AppError {
-
 	dao := DAOFromContext(ctx)
 	gem, _ := GEMFromContext(ctx)
 	gem.Action = "access"
@@ -40,6 +39,8 @@ func (h AppServer) getObjectByPath(ctx context.Context, w http.ResponseWriter, r
 	path, _ = url.PathUnescape(path)
 	pathParts := strings.Split(path, "/")
 	user, _ := UserFromContext(ctx)
+	groupobjects := false
+	groupname := ""
 	var targetObject models.ODObject
 	pagingRequest := protocol.PagingRequest{PageSize: 1}
 	pagingRequest.FilterSettings = []protocol.FilterSetting{}
@@ -47,39 +48,83 @@ func (h AppServer) getObjectByPath(ctx context.Context, w http.ResponseWriter, r
 	for i, part := range pathParts {
 		if len(part) > 0 {
 			if i == 0 {
-				resultset, err := dao.GetRootObjectsByUser(user, mapping.MapPagingRequestToDAOPagingRequest(&pagingRequest))
-				if err != nil {
-					herr := NewAppError(404, err, "Error finding match at root")
-					h.publishError(gem, herr)
-					return herr
-				}
-				if resultset.PageCount > 0 {
-					targetObject = resultset.Objects[0]
+				if strings.ToLower(part) != "groupobjects" {
+					resultset, err := dao.GetRootObjectsByUser(user, mapping.MapPagingRequestToDAOPagingRequest(&pagingRequest))
+					if err != nil {
+						herr := NewAppError(404, err, "Error finding match at root")
+						h.publishError(gem, herr)
+						return herr
+					}
+					if resultset.PageCount > 0 {
+						targetObject = resultset.Objects[0]
+					} else {
+						herr := NewAppError(404, err, "No matching objects at root")
+						h.publishError(gem, herr)
+						return herr
+					}
 				} else {
-					herr := NewAppError(404, err, "No matching objects at root")
-					h.publishError(gem, herr)
-					return herr
+					// NOTE: If a user creates a file/folder named 'groupobjects', it wont get served by this as it will be treated as requesting groups
+					groupobjects = true
 				}
 			} else {
-				partFilterSettings := protocol.FilterSetting{FilterField: "name", Condition: "equals", Expression: pathParts[i]}
-				pagingRequest.FilterSettings[0] = partFilterSettings
-				resultset, err := dao.GetChildObjectsByUser(user, mapping.MapPagingRequestToDAOPagingRequest(&pagingRequest), targetObject)
-				if err != nil {
-					herr := NewAppError(404, err, "Error finding match at folder")
-					h.publishError(gem, herr)
-					return herr
-				}
-				if resultset.PageCount > 0 {
-					targetObject = resultset.Objects[0]
+				if groupobjects {
+					switch i {
+					case 1:
+						groupname = part
+					case 2:
+						partFilterSettings := protocol.FilterSetting{FilterField: "name", Condition: "equals", Expression: part}
+						pagingRequest.FilterSettings[0] = partFilterSettings
+						resultset, err := dao.GetRootObjectsByGroup(groupname, user, mapping.MapPagingRequestToDAOPagingRequest(&pagingRequest))
+						if err != nil {
+							herr := NewAppError(404, err, "Error finding match at groupfolder")
+							h.publishError(gem, herr)
+							return herr
+						}
+						if resultset.PageCount > 0 {
+							targetObject = resultset.Objects[0]
+							// flip back to false as we're no longer at the group root
+							groupobjects = false
+						} else {
+							herr := NewAppError(404, err, "No matching objects in groupfolder")
+							h.publishError(gem, herr)
+							return herr
+						}
+					}
 				} else {
-					herr := NewAppError(404, err, "No matching objects in folder")
-					h.publishError(gem, herr)
-					return herr
+					partFilterSettings := protocol.FilterSetting{FilterField: "name", Condition: "equals", Expression: part}
+					pagingRequest.FilterSettings[0] = partFilterSettings
+					resultset, err := dao.GetChildObjectsByUser(user, mapping.MapPagingRequestToDAOPagingRequest(&pagingRequest), targetObject)
+					if err != nil {
+						herr := NewAppError(404, err, "Error finding match at folder")
+						h.publishError(gem, herr)
+						return herr
+					}
+					if resultset.PageCount > 0 {
+						targetObject = resultset.Objects[0]
+					} else {
+						herr := NewAppError(404, err, "No matching objects in folder")
+						h.publishError(gem, herr)
+						return herr
+					}
 				}
 			}
 		} else {
-			// path ended with /, so this is a folder
-			//captured := make(map[string]string)
+			// path part is empty, effectively ended with /, so this is a folder
+			if !groupobjects {
+				captured["objectId"] = hex.EncodeToString(targetObject.ID)
+				return h.listObjects(context.WithValue(ctx, CaptureGroupsVal, captured), w, r)
+			}
+			// still focused on groups
+			if len(groupname) == 0 {
+				// list group info for user
+				return h.listMyGroupsWithObjects(ctx, w, r)
+			}
+			if i == 2 {
+				// list objects at root for group
+				captured["groupName"] = groupname
+				h.listGroupObjects(context.WithValue(ctx, CaptureGroupsVal, captured), w, r)
+			}
+			// in a subfolder for group, performs as normal list
 			captured["objectId"] = hex.EncodeToString(targetObject.ID)
 			return h.listObjects(context.WithValue(ctx, CaptureGroupsVal, captured), w, r)
 		}